Overview heatmap plotting

TODO: params need to be described

Usage

funky_heatmap(
  data,
  column_info,
  row_info,
  palettes,
  column_groups = NULL,
  row_groups = NULL,
  scale_column = TRUE,
  add_abc = TRUE,
  col_annot_offset = 3,
  row_annot_offset = 0.5,
  removed_methods = NULL
)

Arguments

data

data

column_info

A data frame describing the columns of data. This data frame should contain the following columns:

id (character): The corresponding column name in data.

name (character): A label for the column. If NA or "", no label will be plotted.

group (character): The group of the columns. If all are NA, the columns will not be split up into groups.

geom (character): The geom of the column. Must be one of: circle, rect, bar, pie, or text.

palette (character): Which palette to colour the geom by.

options (list): Column specific options. The content of the list will depend on the geom. Options are:

width: Custom width for this column (default: 1).

overlay: Whether to overlay this column over the previous column. If so, the width of that column will be inherited.

legend: Whether or not to add a legend for this column.

hjust, vjust, size: see ggplot2::geom_text.

label (geom = "text"): Which column to use as a label.

hjust (geom = "bar"): Horizontal alignment of the bar, must be between [0,1].

row_info

A data frame describing the rows of data. This data should contain the following columns:`

id (character): The corresponding row name in data.

group (character): The group of the row. If all are NA, the rows will not be split up into groups.

palettes

e

column_groups

b

row_groups

d

scale_column

f

add_abc

Whether or not to add subfigure labels to the different columns groups.

col_annot_offset

How much the column annotation will be offset by.

row_annot_offset

How much the column annotation will be offset by.

removed_methods

Which methods to not show in the rows