Summary results from the scIB project

This dataset was generated by running the data-raw/scib_summary.R script. This script downloads data for the RNA tasks from the scIB reproducibility repository and summarises it similarly (but not exactly like) what was done in the original paper to give a final ranking of the top performing methods. It is used in the scIB vignette (vignette("scIB")) to reproduce the overall summary figure in Luecken et al. 2021.

Usage

scib_summary

Format

An object of class tbl_df (inherits from tbl, data.frame) with 20 rows and 27 columns.

References

scib_Luecken2021funkyheatmap