Defines parameters for positioning in a plot.

This function sets parameters that affect positioning within a plot, such as row and column dimensions, annotation details, and the expansion directions of the plot.

Usage

position_arguments(
  row_height = 1,
  row_space = 0.1,
  row_bigspace = 0.5,
  col_width = 1,
  col_space = 0.1,
  col_bigspace = 0.5,
  col_annot_offset = 3,
  col_annot_angle = 30,
  expand_xmin = 0,
  expand_xmax = 2,
  expand_ymin = 0,
  expand_ymax = 0
)

Arguments

row_height

The height of the rows.

row_space

The space between rows.

row_bigspace

The large space between row groups.

col_width

The width of the columns.

col_space

The space between columns.

col_bigspace

The large space between column groups.

col_annot_offset

How much the column annotation will be offset by.

col_annot_angle

The angle of the column annotation labels.

expand_xmin

The minimum expansion of the plot in the x direction.

expand_xmax

The maximum expansion of the plot in the x direction.

expand_ymin

The minimum expansion of the plot in the y direction.

expand_ymax

The maximum expansion of the plot in the y direction.

Value

A list of plot positioning parameters.

Examples

position_arguments(row_height = 1.2, col_width = 1.5, expand_xmax = 3)
#> $row_height
#> [1] 1.2
#> 
#> $row_space
#> [1] 0.1
#> 
#> $row_bigspace
#> [1] 0.5
#> 
#> $col_width
#> [1] 1.5
#> 
#> $col_space
#> [1] 0.1
#> 
#> $col_bigspace
#> [1] 0.5
#> 
#> $col_annot_offset
#> [1] 3
#> 
#> $col_annot_angle
#> [1] 30
#> 
#> $expand_xmin
#> [1] 0
#> 
#> $expand_xmax
#> [1] 3
#> 
#> $expand_ymin
#> [1] 0
#> 
#> $expand_ymax
#> [1] 0
#>